Improve class name
