--- Sources/Suspend.swift
@@ -1,8 +1,0 @@
-//
-//  Suspend.swift
-//  UI Queue
-//
-//  Created by srea on 2021/12/12.
-//
-
-import Foundation
