+++ Sources/UIPresentCoordinator.swift
@@ -76,9 +76,6 @@ public final class UIPresentCoordinator: ObservableObject, UIPresentCoordinatabl
     }
 
     @objc func windowDidBecomeHidden(notification: Notification) {
-        guard let window = notification.object as? UIWindow else {
-            return
-        }
         dismissed()
     }
 
@@ -185,54 +182,5 @@ extension UIPresentCoordinator {
             fatalError()
         }
         return ui
-    }
-}
-
-
-public protocol InterruptSuppression {
-    var classNames: [AnyClass] { get }
-}
-
-extension InterruptSuppression {
-    func addClassIfAvailable(list: inout [AnyClass], className: String) {
-        guard let classType = NSClassFromString(className) else {
-            return
-        }
-        list.append(classType)
-    }
-}
-
-
-public extension UIPresentCoordinator {
-
-    struct SystemAlertInterruption: InterruptSuppression {
-
-        public var classNames: [AnyClass] = []
-        
-        public init() {
-            addClassIfAvailable(list: &classNames, className: "UIAlertController")
-            addClassIfAvailable(list: &classNames, className: "SwiftUI.PlatformAlertController")
-            addClassIfAvailable(list: &classNames, className: "SKStoreReviewPresentationWindow")
-        }
-    }
-
-    struct FirebaseInAppMessagingInterruption: InterruptSuppression {
-
-        public var classNames: [AnyClass] = []
-
-        public init() {
-            addClassIfAvailable(list: &classNames, className: "FIRIAMImageOnlyViewController")
-            addClassIfAvailable(list: &classNames, className: "FIRIAMBannerViewController")
-            addClassIfAvailable(list: &classNames, className: "FIRIAMModalViewController")
-            addClassIfAvailable(list: &classNames, className: "FIRIAMCardViewController")
-        }
-    }
-    
-    struct CustomClassInterruption: InterruptSuppression {
-        public var classNames: [AnyClass] = []
-        
-        public init(objects: [AnyClass]) {
-            classNames.append(contentsOf: objects)
-        }
     }
 }
